# Cache select facets with zero/one selected options