--- a/app/views/finders/_select_facet.html.erb
+++ b/app/views/finders/_select_facet.html.erb
@@ -1,10 +1,12 @@
-<% close_facet = select_facet.unselected? && ((select_facet_counter > 0) || select_facet.close_facet?) %>
-<%= render partial: 'components/option-select', locals: {
-  :key => select_facet.key,
-  :title => select_facet.name,
-  :aria_controls_id => "js-search-results-info",
-  :options_container_id => select_facet.key,
-  :options => select_facet.options("js-search-results-info", select_facet.key),
-  :closed_on_load => close_facet,
-}
-%>
+<% cache_if select_facet.cacheable?, select_facet.cache_key do %>
+  <% close_facet = select_facet.unselected? && ((select_facet_counter > 0) || select_facet.close_facet?) %>
+  <%= render partial: 'components/option-select', locals: {
+    :key => select_facet.key,
+    :title => select_facet.name,
+    :aria_controls_id => "js-search-results-info",
+    :options_container_id => select_facet.key,
+    :options => select_facet.options("js-search-results-info", select_facet.key),
+    :closed_on_load => close_facet,
+  }
+  %>
+<% end %>